+++ packages/flame/src/Alert/Alert.tsx
@@ -1,58 +1,19 @@
 import * as React from 'react';
-import styled from '@emotion/styled';
-import { transparentize } from 'polished';
-import { space, variant, SpaceProps } from 'styled-system';
-import { themeGet } from '@styled-system/theme-get';
+import { SpaceProps } from 'styled-system';
+import css from '@styled-system/css';
+
+import { CloseButton } from './CloseButton';
 
 import { Flex, Box } from '../Core';
 import { Text } from '../Text';
 
-const alertStyles = variant({
-  key: 'alertVariants',
-  prop: 'type',
-});
-
-const AlertWrapper = styled('div')<{ type: string }>`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.06), 0 3px 6px 0 rgba(0, 0, 0, 0.03),
-    0 1px 2px 0 rgba(0, 0, 0, 0.1);
-  border-top: 4px solid;
-  border-radius: ${themeGet('radii.radius-2')};
-  padding: ${themeGet('space.2')} ${themeGet('space.3')};
-  ${space}
-  ${alertStyles}
-`;
-
-AlertWrapper.defaultProps = {
-  type: 'info',
-};
-
-const CloseButton = styled('button')`
-  font-size: ${themeGet('fontSizes.text')};
-  color: ${themeGet('colors.textHeading')};
-  background-color: ${props => transparentize(0.9, themeGet('colors.textHeading', '#000')(props))};
-  border-radius: 50%;
-  border: none;
-  cursor: pointer;
-  width: 1em;
-  height: 1em;
-  padding: 0;
-  position: relative;
-
-  &:focus {
-    outline: none;
-  }
-`;
-
-CloseButton.displayName = 'CloseButton';
+type AlertTypes = 'info' | 'success' | 'warning' | 'danger' | string;
 
 export interface AlertProps {
   /** CSS class name */
   className?: string;
   /** Enum for preset Alert types */
-  type?: 'info' | 'success' | 'warning' | 'danger' | string;
+  type?: AlertTypes;
   /** Function called when Close button is tapped */
   onClose?: Function;
   /** Whether a Close button appears */
@@ -74,6 +35,13 @@ export const Alert: React.FC<AlertProps & SpaceProps> = ({
   icon = null,
   ...restProps
 }) => {
+  if (icon) {
+    console.warn(
+      'Warning: Starting from next major version, you will no longer be able to add a prop icon.',
+      'Icons will be automatically be injected based on the Alert type',
+    );
+  }
+
   const [isHidden, setIsHidden] = React.useState(false);
 
   const handleClose = (e: React.MouseEvent<HTMLElement>) => {
@@ -84,14 +52,26 @@ export const Alert: React.FC<AlertProps & SpaceProps> = ({
   if (isHidden) return null;
 
   return (
-    <AlertWrapper type={type} {...restProps}>
-      <Flex flex="1" alignItems="flex-start">
-        {icon && (
-          <Flex alignItems="center" pr={2} pt="2px">
+    <Box
+      css={css({
+        display: 'flex',
+        justifyContent: 'space-between',
+        alignItems: 'flex-start',
+        boxShadow:
+          '0 0 0 1px rgba(0, 0, 0, 0.06), 0 3px 6px 0 rgba(0, 0, 0, 0.03), 0 1px 2px 0 rgba(0, 0, 0, 0.1)',
+        borderTop: '4px solid',
+        borderRadius: 'radius-2',
+        px: 2,
+        py: 3,
+        variant: `alertVariants.${type}`,
+      })}
+      {...restProps}
+    >
+      <Flex flex="1">
+        <Box flex="1" css={css({ position: 'relative', pl: icon ? 4 : 1 })}>
+          <Flex className="fl-alert__icon" css={{ position: 'absolute', left: '0px', top: '2px' }}>
             {icon}
           </Flex>
-        )}
-        <Box flex="1">
           {title && (
             <Text color="textHeading" fontWeight="bold" fontSize="text" mt={0} mr={0} mb={1} ml={0}>
               {title}
@@ -100,31 +80,7 @@ export const Alert: React.FC<AlertProps & SpaceProps> = ({
           <Box fontSize="text-s">{children}</Box>
         </Box>
       </Flex>
-      {!noCloseBtn && (
-        <CloseButton type="button" onClick={handleClose}>
-          <svg
-            xmlns="http://www.w3.org/2000/svg"
-            viewBox="0 0 8 8"
-            css={{
-              width: '0.5em',
-              height: '0.5em',
-              fill: 'black',
-              position: 'absolute',
-              top: '0.25em',
-              left: '0.25em',
-              right: '0',
-              bottom: '0',
-            }}
-          >
-            <g fillRule="evenodd" transform="translate(-4 -4)">
-              <path
-                fillOpacity=".5"
-                d="M9.414 8l2.122-2.121a1 1 0 1 0-1.415-1.415L8 6.586 5.879 4.464A1 1 0 0 0 4.464 5.88L6.586 8l-2.122 2.121a1 1 0 0 0 1.415 1.415L8 9.414l2.121 2.122a1 1 0 0 0 1.415-1.415L9.414 8z"
-              />
-            </g>
-          </svg>
-        </CloseButton>
-      )}
-    </AlertWrapper>
+      {!noCloseBtn && <CloseButton onClick={handleClose} />}
+    </Box>
   );
 };